requirements.txt
django==1.9.5
mysqlclient==1.3.7
python-dateutil